requirements.txt
azure-cognitiveservices-speech==1.18.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
ibm-cloud-sdk-core==3.11.3
ibm-watson==5.2.3
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
websocket-client==1.1.0
Werkzeug==2.0.1
